requirements.txt
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.5
async-lru==2.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==1.11.1
beautifulsoup4==4.10.0
bleach==6.0.0
blinker==1.6.2
boltons==21.0.0
cachetools==5.3.1
caer==2.0.8
certifi==2023.5.7
cffi==1.15.0
chardet==5.0.0
charset-normalizer==3.2.0
cheroot==9.0.0
click==8.1.4
click-log==0.4.0
click-spinner==0.1.10
colorama==0.4.6
colorlog==6.7.0
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.7.1
cycler==0.12.1
dacite==1.6.0
datacommons==1.4.3
datacommons-pandas==0.0.3
datapane==0.16.5
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dnspython==2.3.0
docutils==0.20.1
dominate==2.7.0
dotty-dict==1.3.1
duckdb==0.8.1
duckdb-engine==0.9.0
dulwich==0.20.50
email-validator==2.0.0.post2
entrypoints==0.4
exceptiongroup==1.1.2
executing==0.8.3
face==22.0.0
fastapi==0.100.0
fastjsonschema==2.17.1
filelock==3.12.4
fonttools==4.43.1
fqdn==1.5.1
fsspec==2023.9.2
furl==2.1.3
gitdb==4.0.10
GitPython==3.1.31
glom==22.1.0
gotrue==1.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.10.0
imutils==0.5.4
invoke==1.7.3
ipykernel==6.10.0
ipynbname==2021.3.2
ipython==8.2.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
ipywidgets==7.7.0
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jaraco.functools==3.8.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupysql==0.7.9
jupyter==1.0.0
jupyter-console==6.4.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==7.4.9
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
jupyterlab_server==2.23.0
keyring==24.2.0
kiwisolver==1.4.5
lxml==4.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.3
mdurl==0.1.2
micawber==0.5.4
mimesis==11.1.0
mistune==0.8.4
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
multimethod==1.9.1
munch==2.5.0
mypy==1.6.0
mypy-extensions==1.0.0
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.9.0
nest-asyncio==1.5.4
networkx==3.1
notebook==6.4.10
notebook_shim==0.2.3
numpy==1.25.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
orderedmultidict==1.0.1
orjson==3.9.2
overrides==7.3.1
packaging==23.1
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.8.0
ploomber-core==0.2.13
plotly==5.15.0
postgrest==0.10.6
posthog==2.2.0
premailer==3.10.0
prettytable==3.8.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==4.23.4
psutil==5.9.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.2
pydantic-extra-types==2.0.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyngrok==5.2.3
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-gitlab==3.15.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-semantic-release==7.33.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.3.0
QtPy==2.0.1
readme-renderer==40.0
realtime==1.0.0
referencing==0.29.1
requests==2.31.0
requests-toolbelt==1.0.0
resend==0.5.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.8.10
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.1
SQLAlchemy==2.0.17
sqlglot==16.7.4
sqlparse==0.4.4
stack-data==0.2.0
starlette==0.27.0
storage3==0.5.2
streamlit==1.27.2
streamlit-aggrid==0.3.4.post3
streamlit-card==0.0.61
StrEnum==0.4.15
stringcase==1.2.0
supabase==1.0.3
supafunc==0.2.2
sympy==1.12
tabulate==0.8.10
tenacity==8.2.2
terminado==0.13.3
testpath==0.6.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
twine==3.8.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ujson==5.8.0
uri-template==1.3.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
vega-datasets==0.9.0
watchfiles==0.19.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==10.4
widgetsnbextension==3.6.0
xlrd==2.0.1
XlsxWriter==3.1.2
yagmail==0.15.293
zipp==3.16.0
